fix: false positive
`string.split('\n')` and more where matched. The
commit fixes that problem by requiring that
`Possbility1` is not preceeded by a letter.
Special characters like `$`, `_` etc. are still
matched.

The fix has one downside:
A match must be preceeded by a character now.
Matching the first line and first column of a file,
does not work anymore. This was deemed as
rare edge case and is therefore ignored for now.