fix: writeResources after update on resources
An effect has been used to react to changes of `resources`. Resources is a store though. Listening to all changes is not possible, see https://github.com/solidjs/solid/discussions/829. A custom `setResources` has been added to get rid of the `createEffect` call. Not only does the custom setResources function resolve the bug but the UX flaw of writing to the file system implicitly has been resolved as well.