fix: telefunc functions not executing
The middleware `express.text()` has previously been used in a global context with `app.use(express.text())`. Using the middleware in a global context is not desired as the middleware is telefunc specific. Upon moving the middleware in into the telefunc route "/_telefunc", `app.use()` has not been removed. The latter led to a silent drop in executing telefunc functions.