Refactor memoryFs to use `Uint8Array`
As it turns out, there's actually a significant perfomance hit when
storing binary data (or any data for that matter) as as a string. In
short, because JS uses UTF-16 to encode strings, each 'byte' we store
actually takes up two bytes in memory. Genius.

The result of this is that every filesystem operation takes exactly
twice as long, since twice the amount of data must be read/written. The
simple change from storing data as `Uint8Array` vs `string` means a
`git.clone` operation that used to take 20 seconds now takes about 10.

If you're reading this, and happen to be writing a programing language,
please use UTF-8 like everybody else. Thank you.