Fix import path issue in sdk-js for Windows (#821)
This commit addresses an issue in the inlang sdk-js library where the import function was used to load the svelte.config.js file, resulting in a string that starts with the pattern "c:" on Windows systems. However, the import function expects a path beginning with "file://". This caused problems when starting the SvelteKit development environment or building for production.