refactor: Update openInEditor command title and add branch parameter
The openInEditor command in the Sherlock IDE extension has been updated to change the title to "Sherlock: Open in Fink". Additionally, a new branch parameter has been added to the command callback function, allowing users to specify a branch when opening the editor. This change improves the usability and flexibility of the command.