feat: Add parser for message reference matchers in sap-ui5 IDE extension
This commit adds a new file, messageReferenceMatchers.ts, to the sap-ui5 IDE extension. The file contains a parser that can match different types of message reference patterns in TypeScript, JavaScript, JSON, and XML files. This parser improves the functionality of the IDE extension by allowing it to parse and replace message references more accurately.