implement mutation handler special rules for version updates.
This is a tricky one. What does the developer expect when updating a versions change set id?

We dont know. Hardcoding a heuristic is implicit, hard to reason about, and likely breaks a few use cases.

The solution is to "do nothing" and let the calling code handle edge creation etc.