implement xFilter in virtual table (#3633)
Huge compliment goes out to @martin-lysk who is the actual author of
this PR.

The xFilter method was so painful to implement. Dangling pointers,
SQLite to JS and back to SQLite conversions, no documentation, and yet
he got it done.

This PR is a follow-up of https://github.com/opral/monorepo/pull/3620
that fixes merge conflicts and cleans up a few pieces in the vtable
implementation.

Co-authored-by: Martin Lysk
<113943358+martin-lysk@users.noreply.github.com>